refactor: split daemon/server.ts into auth-manager, config-watcher, ipc-handler modules
